+++ app/ante/ante.go
@@ -11,6 +11,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	authante "github.com/cosmos/cosmos-sdk/x/auth/ante"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	vestingtypes "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
 
 	"github.com/evmos/ethermint/crypto/ethsecp256k1"
 )
@@ -34,6 +35,21 @@ func NewAnteHandler(options HandlerOptions) (sdk.AnteHandler, error) {
 		var anteHandler sdk.AnteHandler
 
 		defer Recover(ctx.Logger(), &err)
+
+		// disable vesting message types in check-tx mode
+		if ctx.IsCheckTx() {
+			for _, msg := range tx.GetMsgs() {
+				switch msg.(type) {
+				case *vestingtypes.MsgCreateVestingAccount,
+					*vestingtypes.MsgCreatePeriodicVestingAccount,
+					*vestingtypes.MsgCreatePermanentLockedAccount:
+					return ctx, sdkerrors.Wrapf(
+						sdkerrors.ErrInvalidRequest,
+						"vesting messages are not supported",
+					)
+				}
+			}
+		}
 
 		txWithExtensions, ok := tx.(authante.HasExtensionOptionsTx)
 		if ok {